Improve interface naming: Renames projects -> InstancesApi
